Add files via upload
The AutomatedGUI is a automatic way of initializing the graphical user interface components of a program such as buttons, text fields, and labels. This is a class I developed and used several times over and over as it simplifies the process of creating a GUI window. Rather than having to explicitly define each component and set its attributes, this method allows this initialization to be done in an automatic and efficient way using an array of objects.