Add files via upload
This is a class I developed that contains several methods regarding objects of the time class. The algorithm I developed traverses an abstract data structure (LinkedList) and looks at the starting time of these objects (or "events"). It then determines which of the objects in the structure is starting in the least amount of time. This file was an essential part of the SportsEX application that can be found here as well.