Add files via upload
This is a class that is used to create LinkedLists, which are data structures used to hold user-entered data. This class is a modified version of the original LinkedList class where each Node can hold objects of a custom-defined class known as "Event" rather than being limited primitive data types such as int or double. This file was also a major part of the SportsEX application available here.